WIP.
No real summary here, just continuing to refactor this into a generally-usable library.

I will say the 3x effort to produce something 'reusable' is really getting borne out
here. The core algorithms aren't changing much and are quite simple, but wrapping this
all in something reusable and extensible is taking longer than it did to write it
in the first place for sure.